Move most things not critical to parsing/lexing to an extras module.
