+++ src/error.rs
@@ -10,11 +10,11 @@ pub enum JsonPopError<'a> {
     Parse(crate::parser::ParseError<'a>),
     Io(std::io::Error),
     /// This type is a never variation unless the testsuite is being run.
-    TestError(crate::test_utils::TestError<'a>),
+    TestError(crate::extra::test_utils::TestError<'a>),
 }
 
 /// A top level error returned by processes or tests..
-/// It's not bounded by the lifetime of the program
+/// It's not bounded by the lifetime of the string being parsed. 
 /// we should add error codes to these.
 ///
 /// That will be a breaking change.
@@ -38,8 +38,8 @@ impl<'a> From<JsonPopError<'a>> for TopLevelError {
     }
 }
 
-// This error lives inside the the parsers Error type.
-// So it's a sub-error of a parse error.
+/// This error lives inside the the parsers Error type.
+/// So it's a sub-error of a parse error.
 #[derive(Debug)]
 pub enum CompilationError {
     LexicalError { range: Range<usize> },
